requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Werkzeug==2.3.7
pymysqlpool==0.3.4
requests==2.31.0
python-dotenv==1.0.0
